Removed all calls to malloc/free. Everything now uses a fixed memory budget allocated at startup. A lot of file routines and startup procedures were rearranged to accomodate this change. Also, this doesn't remove the CRT memory functions entirely as they are implictly called by other C standard library calls as well as X11 and linux system calls. Total removal would require overriding them with similar calls that route to our budgeted memory and making those thread-safe, but that's too much effort for now.

+++ random.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include "sized_types.h"
+
 namespace random {
 
-void seed(unsigned long seed);
-unsigned long generate();
+void seed(u32 seed);
+u32 generate();
 int int_range(int min, int max);
 float float_range(float min, float max);
 double double_range(double min, double max);
